messaging: remove KUBE signaling client (WebsocketSignalManager/SignalClient)
Delete the deprecated KUBE websocket signaling stack — WebsocketSignalManager, SignalClient,
SignalRPCClient, SignalLocalState, their monitors, and the SignalClientMethods interface. Its test
harness (@dxos/signal) was already removed and edge signaling (EdgeSignalManager) is the only real
transport.

Rewire the non-edge fallback in the services host, worker runtime, and local client services to
MemorySignalManager (isolated, no cross-process signaling), and drop the now-dead KUBE-only telemetry
plumbing (signalMetadata callback, signalTelemetryEnabled). Test builders (network-manager,
client-services, client, blade-runner) and the two skipped browser tests move to MemorySignalManager /
the default memory provider; the skipped KUBE load test is deleted.
