+++ packages/core/mesh/messaging/src/index.ts
@@ -3,6 +3,5 @@
 //
 
 export * from './messenger';
-export * from './signal-client';
 export * from './signal-manager';
 export * from './signal-methods';
