--- packages/core/mesh/messaging/src/signal-client/index.ts
@@ -1,5 +1,0 @@
-//
-// Copyright 2022 DXOS.org
-//
-
-export * from './signal-client';
