+++ packages/core/mesh/messaging/src/signal-manager/index.ts
@@ -4,6 +4,5 @@
 
 export * from './memory-signal-manager';
 export * from './signal-manager';
-export * from './websocket-signal-manager';
 export * from './edge-signal-manager';
 export * from './utils';
