+++ packages/sdk/client-services/src/packlets/space/space-manager.browser.test.ts
@@ -6,18 +6,12 @@ import { describe, onTestFinished, test } from 'vitest';
 
 import { createStorage } from '@dxos/random-access-storage';
 
-import { TestAgentBuilder, WebsocketNetworkManagerProvider } from './testing';
-
-// TODO(burdon): Config.
-// Signal server will be started by the setup script.// Signal server will be started by the setup script.
-const port = process.env.SIGNAL_PORT ?? 4000;
-const SIGNAL_URL = `ws://localhost:${port}/.well-known/dx/signal`;
+import { TestAgentBuilder } from './testing';
 
 describe('space-manager', () => {
   test.skip('invitations', async () => {
     const builder = new TestAgentBuilder({
       storage: createStorage(),
-      networkManagerProvider: WebsocketNetworkManagerProvider(SIGNAL_URL),
     });
     onTestFinished(async () => {
       await builder.close();
