+++ packages/sdk/client-services/src/packlets/space/testing/test-agent-builder.ts
@@ -7,8 +7,8 @@ import { CredentialGenerator } from '@dxos/credentials';
 import { type FeedStore } from '@dxos/feed-store';
 import { type Keyring } from '@dxos/keyring';
 import { PublicKey } from '@dxos/keys';
-import { MemorySignalManager, MemorySignalManagerContext, WebsocketSignalManager } from '@dxos/messaging';
-import { MemoryTransportFactory, SwarmNetworkManager, createRtcTransportFactory } from '@dxos/network-manager';
+import { MemorySignalManager, MemorySignalManagerContext } from '@dxos/messaging';
+import { MemoryTransportFactory, SwarmNetworkManager } from '@dxos/network-manager';
 import { type FeedMessage } from '@dxos/protocols/proto/dxos/echo/feed';
 import { type SpaceMetadata } from '@dxos/protocols/proto/dxos/echo/metadata';
 import { AdmittedFeed } from '@dxos/protocols/proto/dxos/halo/credentials';
@@ -29,14 +29,6 @@ export const MemoryNetworkManagerProvider =
     new SwarmNetworkManager({
       signalManager: new MemorySignalManager(signalContext),
       transportFactory: MemoryTransportFactory,
-    });
-
-export const WebsocketNetworkManagerProvider =
-  (signalUrl: string): NetworkManagerProvider =>
-  () =>
-    new SwarmNetworkManager({
-      signalManager: new WebsocketSignalManager([{ server: signalUrl }]),
-      transportFactory: createRtcTransportFactory(),
     });
 
 export type TestAgentBuilderOptions = {
