+++ packages/sdk/client/src/services/local-client-services.ts
@@ -69,16 +69,7 @@ export const fromHost = async (
   config = new Config(),
   params?: LocalClientServicesParams,
 ): Promise<ClientServicesProvider> => {
-  const observabilityGroup = config.get('runtime.client.observabilityGroup');
-  const signalTelemetryEnabled = config.get('runtime.client.signalTelemetryEnabled');
-  const networking = await setupNetworking(config, {}, () =>
-    signalTelemetryEnabled
-      ? {
-          ...services.signalMetadataTags,
-          ...(observabilityGroup ? { group: observabilityGroup } : {}),
-        }
-      : {},
-  );
+  const networking = await setupNetworking(config, {});
 
   const services = new LocalClientServices({ config, ...networking, ...params });
   return services;
@@ -91,21 +82,20 @@ export const fromHost = async (
 const setupNetworking = async (
   config: Config,
   options: Partial<SwarmNetworkManagerOptions> = {},
-  signalMetadata?: () => void,
 ): Promise<{
   signalManager?: SignalManager;
   transportFactory: TransportFactory;
 }> => {
-  const { MemorySignalManager, MemorySignalManagerContext, WebsocketSignalManager } = await import('@dxos/messaging');
+  const { MemorySignalManager, MemorySignalManagerContext } = await import('@dxos/messaging');
   const { createRtcTransportFactory, MemoryTransportFactory } = await import('@dxos/network-manager');
 
   const signals = config.get('runtime.services.signaling');
   const edgeFeatures = config.get('runtime.client.edgeFeatures');
   if (signals || edgeFeatures?.signaling) {
     const {
-      signalManager = edgeFeatures?.signaling || !signals
-        ? undefined // EdgeSignalManager needs EdgeConnection and will be created in service-host
-        : new WebsocketSignalManager(signals, signalMetadata),
+      // EdgeSignalManager needs an EdgeConnection and is created in the services host; without edge
+      // signaling fall back to an isolated in-memory manager (KUBE `WebsocketSignalManager` removed).
+      signalManager = edgeFeatures?.signaling ? undefined : new MemorySignalManager(new MemorySignalManagerContext()),
       // TODO(wittjosiah): P2P networking causes seg fault in bun currently.
       transportFactory = isBun()
         ? MemoryTransportFactory
